+++ packages/web3-core-method/src/index.js
@@ -23,7 +23,6 @@
 
 'use strict';
 
-var _ = require('underscore');
 var errors = require('web3-core-helpers').errors;
 var formatters = require('web3-core-helpers').formatters;
 var utils = require('web3-utils');
@@ -102,7 +101,7 @@ Method.prototype.attachToObject = function (obj) {
  * @return {String} name of jsonrpc method
  */
 Method.prototype.getCall = function (args) {
-    return _.isFunction(this.call) ? this.call(args) : this.call;
+    return typeof this.call === 'function' ? this.call(args) : this.call;
 };
 
 /**
@@ -113,7 +112,7 @@ Method.prototype.getCall = function (args) {
  * @return {Function|Null} callback, if exists
  */
 Method.prototype.extractCallback = function (args) {
-    if (_.isFunction(args[args.length - 1])) {
+    if (typeof (args[args.length - 1]) === 'function') {
         return args.pop(); // modify the args array!
     }
 };
@@ -161,7 +160,7 @@ Method.prototype.formatInput = function (args) {
 Method.prototype.formatOutput = function (result) {
     var _this = this;
 
-    if (_.isArray(result)) {
+    if (Array.isArray(result)) {
         return result.map(function (res) {
             return _this.outputFormatter && res ? _this.outputFormatter(res) : res;
         });
@@ -180,6 +179,7 @@ Method.prototype.formatOutput = function (result) {
 Method.prototype.toPayload = function (args) {
     var call = this.getCall(args);
     var callback = this.extractCallback(args);
+    
     var params = this.formatInput(args);
     this.validateArgs(params);
 
@@ -206,8 +206,8 @@ Method.prototype._confirmTransaction = function (defer, result, payload) {
         intervalId = null,
         lastBlock = null,
         receiptJSON = '',
-        gasProvided = (_.isObject(payload.params[0]) && payload.params[0].gas) ? payload.params[0].gas : null,
-        isContractDeployment = _.isObject(payload.params[0]) &&
+        gasProvided = ((!!payload.params[0] && typeof payload.params[0] === 'object') && payload.params[0].gas) ? payload.params[0].gas : null,
+        isContractDeployment = (!!payload.params[0] && typeof payload.params[0] === 'object') &&
             payload.params[0].data &&
             payload.params[0].from &&
             !payload.params[0].to,
@@ -258,7 +258,7 @@ Method.prototype._confirmTransaction = function (defer, result, payload) {
     ];
     // attach methods to this._ethereumCall
     var _ethereumCall = {};
-    _.each(_ethereumCalls, function (mthd) {
+    _ethereumCalls.forEach(mthd =>  {
         mthd.attachToObject(_ethereumCall);
         mthd.requestManager = method.requestManager; // assign rather than call setRequestManager()
     });
@@ -591,11 +591,11 @@ var getWallet = function (from, accounts) {
     var wallet = null;
 
     // is index given
-    if (_.isNumber(from)) {
+    if (typeof from === 'number') {
         wallet = accounts.wallet[from];
 
         // is account given
-    } else if (_.isObject(from) && from.address && from.privateKey) {
+    } else if (!!from && typeof from === 'object' && from.address && from.privateKey) {
         wallet = from;
 
         // search in wallet for address
@@ -689,10 +689,10 @@ Method.prototype.buildCall = function () {
         // SENDS the SIGNED SIGNATURE
         var sendSignedTx = function (sign) {
 
-            var signedPayload = _.extend({}, payload, {
+            var signedPayload = { ... payload, 
                 method: 'eth_sendRawTransaction',
                 params: [sign.rawTransaction]
-            });
+            };
 
             method.requestManager.send(signedPayload, sendTxCallback);
         };
@@ -706,29 +706,30 @@ Method.prototype.buildCall = function () {
                 // ETH_SENDTRANSACTION
                 if (payload.method === 'eth_sendTransaction') {
                     var tx = payload.params[0];
-                    wallet = getWallet((_.isObject(tx)) ? tx.from : null, method.accounts);
+                    wallet = getWallet((!!tx && typeof tx === 'object') ? tx.from : null, method.accounts);
 
 
                     // If wallet was found, sign tx, and send using sendRawTransaction
                     if (wallet && wallet.privateKey) {
-                        var txOptions = _.omit(tx, 'from');
+                        var tx = JSON.parse(JSON.stringify(tx));
+                        delete tx.from;
 
-                        if (method.defaultChain && !txOptions.chain) {
-                            txOptions.chain = method.defaultChain;
+                        if (method.defaultChain && !tx.chain) {
+                            tx.chain = method.defaultChain;
                         }
 
-                        if (method.defaultHardfork && !txOptions.hardfork) {
-                            txOptions.hardfork = method.defaultHardfork;
+                        if (method.defaultHardfork && !tx.hardfork) {
+                            tx.hardfork = method.defaultHardfork;
                         }
 
-                        if (method.defaultCommon && !txOptions.common) {
-                            txOptions.common = method.defaultCommon;
+                        if (method.defaultCommon && !tx.common) {
+                            tx.common = method.defaultCommon;
                         }
 
-                        method.accounts.signTransaction(txOptions, wallet.privateKey)
+                        method.accounts.signTransaction(tx, wallet.privateKey)
                             .then(sendSignedTx)
                             .catch(function (err) {
-                                if (_.isFunction(defer.eventEmitter.listeners) && defer.eventEmitter.listeners('error').length) {
+                                if (typeof defer.eventEmitter.listeners === 'function' && defer.eventEmitter.listeners('error').length) {
                                     try {
                                         defer.eventEmitter.emit('error', err);
                                     } catch (err) {
@@ -769,7 +770,7 @@ Method.prototype.buildCall = function () {
         };
 
         // Send the actual transaction
-        if (isSendTx && _.isObject(payload.params[0]) && typeof payload.params[0].gasPrice === 'undefined') {
+        if (isSendTx && !!payload.params[0] && typeof payload.params[0] === 'object' && typeof payload.params[0].gasPrice === 'undefined') {
 
             var getGasPrice = (new Method({
                 name: 'getGasPrice',
@@ -866,7 +867,7 @@ Method.prototype.getRevertReason = function (txOptions, blockNumber) {
  * @returns {Boolean}
  */
 Method.prototype.isRevertReasonString = function (data) {
-    return _.isString(data) && ((data.length - 2) / 2) % 32 === 4 && data.substring(0, 10) === '0x08c379a0';
+    return typeof data === 'string' && ((data.length - 2) / 2) % 32 === 4 && data.substring(0, 10) === '0x08c379a0';
 };
 
 /**